+++ client/src/components/container/_index.scss
@@ -3,4 +3,5 @@
 @forward "./CartSection/CartSection";
 @forward "./SingleProductSection/SingleProductSection";
 @forward "./LoginSection/LoginSection";
+@forward "./SignUpSection/SignUpSection";
 @forward "./ListingSection/ListingSection";
