+++ client/src/components/container/index.js
@@ -2,6 +2,7 @@ import LandingSection from './LandingSection/LandingSection';
 import StockSection from './StockSection/StockSection';
 import SingleProductSection from "./SingleProductSection/SingleProductSection";
 import LoginSection from "./LoginSection/LoginSection";
+import SignUpSection from "./SignUpSection/SignUpSection";
 import ListingSection from "./ListingSection/ListingSection";
 import CartSection from "./CartSection/CartSection";
 
@@ -9,6 +10,7 @@ export {
     LandingSection,
     StockSection,
     LoginSection,
+    SignUpSection,
     SingleProductSection,
     ListingSection,
     CartSection,
