+++ client/src/style/style.css
@@ -707,6 +707,69 @@ button:focus-visible {
   color: var(--clr-darkgray);
 }
 
+#suSection {
+  display: grid;
+  width: 100%;
+  height: 100%;
+  place-items: center;
+}
+#suSection .suContainer {
+  display: grid;
+  width: 40%;
+  height: 98%;
+  padding: 2rem 0;
+  grid-template-rows: 0.25fr 0.75fr;
+  place-items: center;
+  gap: 2rem;
+  border: 1px solid var(--clr-darkgray);
+  border-radius: 18px;
+}
+#suSection .suContainer .suLogoCropped {
+  font-size: 3.75rem;
+  color: var(--clr-white);
+  -webkit-text-stroke: 1px var(--clr-darkgray);
+}
+#suSection .suContainer .suForm {
+  display: grid;
+  gap: 1.5rem;
+  place-items: center;
+}
+#suSection .suContainer .suForm .suBtn {
+  width: 60%;
+  padding: 0.5rem;
+  color: var(--clr-darkgray);
+  font-size: 1.1rem;
+  background: var(--clr-white);
+  border: 1px solid var(--clr-darkgray);
+  border-radius: 8px;
+  outline: none;
+  transition: 0.3s;
+}
+#suSection .suContainer .suForm .suBtn:hover, #suSection .suContainer .suForm .suBtn:focus-visible {
+  color: var(--clr-white);
+  background: var(--clr-cyan);
+  border-color: var(--clr-white);
+  cursor: pointer;
+}
+#suSection .suContainer .suForm .suInput {
+  display: grid;
+  gap: 0.25rem;
+}
+#suSection .suContainer .suForm .suInput > * {
+  font-size: 1rem;
+}
+#suSection .suContainer .suForm .suInput label {
+  font-size: 1.1rem;
+  font-weight: 600;
+}
+#suSection .suContainer .suForm .suInput label span {
+  color: var(--clr-red);
+}
+#suSection .suContainer .suForm .suInput input {
+  width: 100%;
+  padding: 0.25rem;
+}
+
 #listingPage {
   display: flex;
   flex-direction: column;
@@ -750,69 +813,6 @@ button:focus-visible {
 #listingPage .listingSubmitButton {
   height: 30px;
   margin: 30px 0;
-}
-
-#suPage {
-  display: grid;
-  width: 100%;
-  height: 100%;
-  place-items: center;
-}
-#suPage .suContainer {
-  display: grid;
-  width: 40%;
-  height: 98%;
-  padding: 2rem 0;
-  grid-template-rows: 0.25fr 0.75fr;
-  place-items: center;
-  gap: 2rem;
-  border: 1px solid var(--clr-darkgray);
-  border-radius: 18px;
-}
-#suPage .suContainer .suLogoCropped {
-  font-size: 3.75rem;
-  color: var(--clr-white);
-  -webkit-text-stroke: 1px var(--clr-darkgray);
-}
-#suPage .suContainer .suForm {
-  display: grid;
-  gap: 1.5rem;
-  place-items: center;
-}
-#suPage .suContainer .suForm .suBtn {
-  width: 60%;
-  padding: 0.5rem;
-  color: var(--clr-darkgray);
-  font-size: 1.1rem;
-  background: var(--clr-white);
-  border: 1px solid var(--clr-darkgray);
-  border-radius: 8px;
-  outline: none;
-  transition: 0.3s;
-}
-#suPage .suContainer .suForm .suBtn:hover, #suPage .suContainer .suForm .suBtn:focus-visible {
-  color: var(--clr-white);
-  background: var(--clr-cyan);
-  border-color: var(--clr-white);
-  cursor: pointer;
-}
-#suPage .suContainer .suForm .suInput {
-  display: grid;
-  gap: 0.25rem;
-}
-#suPage .suContainer .suForm .suInput > * {
-  font-size: 1rem;
-}
-#suPage .suContainer .suForm .suInput label {
-  font-size: 1.1rem;
-  font-weight: 600;
-}
-#suPage .suContainer .suForm .suInput label span {
-  color: var(--clr-red);
-}
-#suPage .suContainer .suForm .suInput input {
-  width: 100%;
-  padding: 0.25rem;
 }
 
 #checkoutPage {
